eWAY Rapid 3.1 gateway
Thanks to @moklett who had a major part in helping me put this together.

This introduces the EwayRapid31 gateway. This is separate from the
EwayRapid gateway because it uses eWAY Rapid 3.1's Direct API, so it
does not make any use of the transparent redirect and makes all direct
API calls.