add 'reference' attr to CreditCard
This adds a reference attribute to CreditCard (credit_card.reference)
that is for setting a special value for that card. A good example is
usage with the Realex gateway.

Realex requires a reference value for credit cards stored because there
are 'payment methods' on 'payer'. Realex uses the reference on the
stored card when making transactions as a token along with the payer
reference.